+++ packages/recursive/node/index.js
@@ -1,3 +1,4 @@
 export { Node }       from './src/Node.js'
 export { Grad }       from './src/Grad.js'
 export { Fold, tabs } from './src/Fold.js'
+export { parsePresm } from './src/parsePresm.js'