+++ packages/recursive/node/test/node.matrix.camp.test.js
@@ -1,4 +1,5 @@
-import { BESQUE, METRO, OCEAN }           from '@palett/presets'
+import { Presm }                          from '@palett/pres'
+import { METRO, OCEAN, SUBTLE }           from '@palett/presets'
 import { COLUMNWISE, POINTWISE, ROWWISE } from '@vect/matrix'
 import { indexed, mapVal }                from '@vect/object-mapper'
 import { test }                           from 'node:test'
@@ -9,13 +10,14 @@ const MATRICES = {
   emptyMatrix: [ [] ],
   oneRow: [ [ 0, 0, 1, 8 ] ],
   oneColumn: [ [ 0 ], [ 9 ], [ 11 ] ],
+  oneColumn2: [ [ 'Imperial' ], [ 'Castle' ], [ 'Feudal' ], [ 'Dark' ] ],
   pos: [ [ 0, 0, 1, 8 ], [ 9, 10, 11, 11 ], [ 11, 13, 14, 14 ] ],
   neg: [ [ 0, 0, -1, -8 ], [ -9, -10, -11, -11 ], [ -11, -13, -14, -14 ] ],
-  mix: [ [ 0, 0, -1, -8 ], [ 9, 10, 11, 11 ], [ -11, -13, -14, -14 ] ]
+  mix: [ [ 0, 0, -1, -8 ], [ 9, 10, 11, 11 ], [ -11, -13, -14, -14 ] ],
 }
 
-test('node matrix', () => {
-  const node = new Node({ fill: ' ', ansi: true, pres: { pos: BESQUE, neg: OCEAN, str: METRO } })
+test('node matrix camp', () => {
+  const node = new Node({ fill: ' ', ansi: true, pres: Presm.build(SUBTLE, OCEAN, METRO) })
   const matrices = mapVal(MATRICES, mat => node.matrix(mat, POINTWISE, 2))
   const rows = mapVal(MATRICES, mat => node.matrix(mat, ROWWISE))
   const columns = mapVal(MATRICES, mat => node.matrix(mat, COLUMNWISE))