+++ packages/recursive/deco-flat/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @spare/deco-flat
 
+## 0.16.10
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/deco-date@0.16.10
+  - @spare/deco-func@0.16.10
+  - @spare/node@0.16.10
+  - @spare/deco-colors@0.16.10
+
+## 0.16.9
+
+### Patch Changes
+
+- update interface
+- Updated dependencies
+  - @spare/deco-date@0.16.9
+  - @spare/deco-func@0.16.9
+  - @spare/node@0.16.9
+  - @spare/deco-colors@0.16.9
+
 ## 0.16.8
 
 ### Patch Changes