+++ packages/recursive/deco-func/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @spare/deco-func
 
+## 0.16.10
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/preset-deco@0.16.10
+  - @spare/translator@0.16.10
+  - @spare/deco-config@0.16.10
+
+## 0.16.9
+
+### Patch Changes
+
+- update interface
+- Updated dependencies
+  - @spare/preset-deco@0.16.9
+  - @spare/deco-config@0.16.9
+  - @spare/translator@0.16.9
+
 ## 0.16.8
 
 ### Patch Changes